cli.output: split up module into subpackage
- Move abstract `Output` base class into `streamlink_cli.output.abc`
  and properly declare abstract methods
- Move `FileOutput` into `streamlink_cli.output.file`
- Move `PlayerOutput` into `streamlink_cli.output.player`
- Update patched module paths in tests
