+++ src/streamlink_cli/output/__init__.py
@@ -1,0 +1,3 @@
+from streamlink_cli.output.abc import Output
+from streamlink_cli.output.file import FileOutput
+from streamlink_cli.output.player import PlayerOutput
