+++ tests/cli/output/test_output.py
@@ -10,7 +10,7 @@ import pytest
 from streamlink_cli.output import FileOutput, PlayerOutput
 
 
-@patch("streamlink_cli.output.stdout")
+@patch("streamlink_cli.output.file.stdout")
 class TestFileOutput(unittest.TestCase):
     @staticmethod
     def subject(filename, fd):
@@ -75,7 +75,7 @@ class TestFileOutput(unittest.TestCase):
         self._test_open(mock_open, mock_stdout)
 
     @pytest.mark.windows_only()
-    @patch("streamlink_cli.output.msvcrt")
+    @patch("streamlink_cli.output.file.msvcrt")
     @patch("builtins.open")
     def test_open_windows(self, mock_open: Mock, mock_msvcrt: Mock, mock_stdout: Mock):
         mock_path = self._test_open(mock_open, mock_stdout)
@@ -91,34 +91,34 @@ class TestPlayerOutput(unittest.TestCase):
         assert PlayerOutput.supported_player("mpv") == "mpv"
         assert PlayerOutput.supported_player("potplayermini.exe") == "potplayer"
 
-    @patch("streamlink_cli.output.os.path.basename", new=ntpath.basename)
+    @patch("streamlink_cli.output.player.os.path.basename", new=ntpath.basename)
     def test_supported_player_win32(self):
         assert PlayerOutput.supported_player("C:\\MPV\\mpv.exe") == "mpv"
         assert PlayerOutput.supported_player("C:\\VLC\\vlc.exe") == "vlc"
         assert PlayerOutput.supported_player("C:\\PotPlayer\\PotPlayerMini64.exe") == "potplayer"
 
-    @patch("streamlink_cli.output.os.path.basename", new=posixpath.basename)
+    @patch("streamlink_cli.output.player.os.path.basename", new=posixpath.basename)
     def test_supported_player_posix(self):
         assert PlayerOutput.supported_player("/usr/bin/mpv") == "mpv"
         assert PlayerOutput.supported_player("/usr/bin/vlc") == "vlc"
 
-    @patch("streamlink_cli.output.os.path.basename", new=ntpath.basename)
+    @patch("streamlink_cli.output.player.os.path.basename", new=ntpath.basename)
     def test_supported_player_args_win32(self):
         assert PlayerOutput.supported_player("C:\\MPV\\mpv.exe --argh") == "mpv"
         assert PlayerOutput.supported_player("C:\\VLC\\vlc.exe --argh") == "vlc"
         assert PlayerOutput.supported_player("C:\\PotPlayer\\PotPlayerMini64.exe --argh") == "potplayer"
 
-    @patch("streamlink_cli.output.os.path.basename", new=posixpath.basename)
+    @patch("streamlink_cli.output.player.os.path.basename", new=posixpath.basename)
     def test_supported_player_args_posix(self):
         assert PlayerOutput.supported_player("/usr/bin/mpv --argh") == "mpv"
         assert PlayerOutput.supported_player("/usr/bin/vlc --argh") == "vlc"
 
-    @patch("streamlink_cli.output.os.path.basename", new=posixpath.basename)
+    @patch("streamlink_cli.output.player.os.path.basename", new=posixpath.basename)
     def test_supported_player_negative_posix(self):
         assert PlayerOutput.supported_player("/usr/bin/xmpvideo") is None
         assert PlayerOutput.supported_player("/usr/bin/echo") is None
 
-    @patch("streamlink_cli.output.os.path.basename", new=ntpath.basename)
+    @patch("streamlink_cli.output.player.os.path.basename", new=ntpath.basename)
     def test_supported_player_negative_win32(self):
         assert PlayerOutput.supported_player("C:\\mpc\\mpc-hd.exe") is None
         assert PlayerOutput.supported_player("C:\\mplayer\\not-vlc.exe") is None
